pyproject.toml
[project]
name = "study43"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi-extra>=0.3.3",
    "lxml>=6.0.2",
    "mariadb>=1.1.14",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "standard>=1.0.3",
    "streamlit>=1.55.0",
]
